fix: restore search query across process death, avoid discount overflow
1. fix: persist search query in SavedStateHandle
   - SearchViewModel.currentQuery now initializes from and writes to
     SavedStateHandle (key "search_query"), matching the existing
     barcode_query pattern already used in this ViewModel
   - On restoration, if a query was saved and no barcode result takes
     priority, queryFlow is primed so the search re-runs automatically
   - Previously: plain MutableStateFlow("") meant a process death (low
     memory, OS reclaim) silently dropped whatever the user had typed
     or last searched, forcing them to retype from scratch

2. fix: avoid Long overflow in CurrencyFormatter.discountPercent()
   - (original - current) * 100 could overflow Long for extreme inputs,
     wrapping to a negative value and silently misrepresenting the
     discount. Switched the multiply/divide to Double arithmetic, which
     saturates instead of wrapping.
   - Added a regression test with Long.MAX_VALUE/2 and /4 inputs.

3. refactor: deduplicate LocalNotificationManager alert builders
   - sendPriceAlert() and sendStockAlert() had near-identical Intent /
     PendingIntent / NotificationCompat.Builder chains that could drift
     if only one was updated. Extracted sendProductNotification() as
     the shared implementation; each caller now only supplies its
     notification ID, title, body, and failure log message.

Addresses HIGH/MEDIUM/LOW findings from codebase improvement scan (Tier 94).

+++ app/src/main/java/com/example/popcoon/core/CurrencyFormatter.kt
@@ -26,10 +26,16 @@ object CurrencyFormatter {
         return "$sign¥${String.format(Locale.US, "%,d", abs(diff))}"
     }
 
-    /** 割引率 (例: "20% OFF") */
+    /**
+     * 割引率 (例: "20% OFF")
+     *
+     * (original - current) * 100 を Long のまま計算すると、両者が Long.MAX_VALUE 近くの
+     * 極端な値の場合にオーバーフローして符号が反転しうる (円価格では非現実的だが、
+     * 将来の多通貨対応や不正な外部データ流入に備えて Double 計算で桁あふれを避ける)。
+     */
     fun discountPercent(original: Long, current: Long): String {
         if (original <= 0) return ""
-        val pct = ((original - current) * 100 / original).toInt()
+        val pct = ((original - current).toDouble() * 100.0 / original.toDouble()).toInt()
         return if (pct > 0) "${pct}% OFF" else ""
     }
